test(refactor): move resources to testclasses in fieldaccesses (#2168)
* Move resources to testclasses in fieldaccesses

* Fix imports: replace '*' with single imports

* Place MyClass in testclasses
