+++ src/test/java/spoon/test/fieldaccesses/FieldAccessTest.java
@@ -30,9 +30,11 @@ import spoon.reflect.visitor.filter.NamedElementFilter;
 import spoon.reflect.visitor.filter.TypeFilter;
 import spoon.test.fieldaccesses.testclasses.B;
 import spoon.test.fieldaccesses.testclasses.Kuu;
+import spoon.test.fieldaccesses.testclasses.Mouse;
 import spoon.test.fieldaccesses.testclasses.Panini;
 import spoon.test.fieldaccesses.testclasses.Pozole;
 import spoon.test.fieldaccesses.testclasses.Tacos;
+import spoon.test.fieldaccesses.testclasses.MyClass;
 import spoon.testing.utils.ModelUtils;
 
 import java.util.List;
@@ -50,7 +52,7 @@ public class FieldAccessTest {
 
 	@Test
 	public void testModelBuildingFieldAccesses() throws Exception {
-		CtType<?> type = build("spoon.test.fieldaccesses", "Mouse");
+		CtType<?> type = build("spoon.test.fieldaccesses.testclasses", "Mouse");
 		assertEquals("Mouse", type.getSimpleName());
 
 		CtMethod<?> meth1 = type.getElements(
@@ -98,7 +100,7 @@ public class FieldAccessTest {
 
 	@Test
 	public void testBCUBug20140402() throws Exception {
-		CtType<?> type = build("spoon.test.fieldaccesses",
+		CtType<?> type = build("spoon.test.fieldaccesses.testclasses",
 				"BCUBug20140402");
 		assertEquals("BCUBug20140402", type.getSimpleName());
 
@@ -140,7 +142,7 @@ public class FieldAccessTest {
 
 	@Test
 	public void testBUG20160112() throws Exception {
-		CtType<?> type = build("spoon.test.fieldaccesses", "BUG20160112");
+		CtType<?> type = build("spoon.test.fieldaccesses.testclasses", "BUG20160112");
 		assertEquals("BUG20160112", type.getSimpleName());
 		CtOperatorAssignment<?, ?> ass = type.getElements(
 				new TypeFilter<CtOperatorAssignment<?, ?>>(CtOperatorAssignment.class)).get(0);
@@ -151,7 +153,7 @@ public class FieldAccessTest {
 
 	@Test
 	public void testTargetedAccessPosition() throws Exception {
-		CtType<?> type = build("spoon.test.fieldaccesses", "TargetedAccessPosition");
+		CtType<?> type = build("spoon.test.fieldaccesses.testclasses", "TargetedAccessPosition");
 		List<CtFieldAccess<?>> vars = type.getElements(
 				new TypeFilter<CtFieldAccess<?>>(CtFieldAccess.class));
 		//vars is [t.ta.ta, t.ta]
@@ -185,7 +187,7 @@ public class FieldAccessTest {
 		assertEquals("LOG", logFieldAccess.getVariable().getSimpleName());
 		assertEquals(MyClass.class, logFieldAccess.getVariable().getDeclaringType().getActualClass());
 
-		String expectedLambda = "() -> {" + System.lineSeparator() + "    spoon.test.fieldaccesses.MyClass.LOG.info(\"bla\");" + System.lineSeparator() + "}";
+		String expectedLambda = "() -> {" + System.lineSeparator() + "    spoon.test.fieldaccesses.testclasses.MyClass.LOG.info(\"bla\");" + System.lineSeparator() + "}";
 		assertEquals(expectedLambda, logFieldAccess.getParent(CtLambda.class).toString());
 	}
